thanh mau, hieu ung khi chet, player nhan dame

+++ Assets/Script/Monsters/Forest/MonsterShoot.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class MonsterShoot : MonoBehaviour
 {
-    public float dame;
+    public int dame;
     fire myfire;
     public GameObject bulletExplosion;
